structure config.py into multiple files
