+++ src/luxonis_ml/scheme/v1/config.py
@@ -1,43 +1,6 @@
 from pydantic import BaseModel
 from typing import Optional, Dict, Any, List, Tuple, Literal, Union, Annotated
-from enum import Enum
-
-# ---- ENUM CLASSES ----
-
-class Platform(Enum):
-    HAILO = "hailo"
-    RVC2 = "rvc2"
-    RVC3 = "rvc3"
-    RVC4 = "rvc4"
-
-class DataType(Enum):
-    INT8 = "int8"
-    UINT8 = "uint8"
-    FLOAT32 = "float32"
-
-class InputType(Enum):
-    RAW = "raw"
-    IMAGE = "image"
-
-class PreprocessingBlockType(Enum): # required only for RVC4
-    MEAN = "mean"
-    SCALE = "scale"
-    REVERSE_CHANNELS = "reverse_channels" #bgr<->rgb
-
-class PostprocessingBlockType(Enum):
-    MISC = "misc" # TODO
-    
-class TaskType(Enum):
-    CLASSIFICATION = "classification" # ?
-    MULTICLASS_CLASSIFICATION = "multiclass_classification" # e.g. male-blue, female-blue, male-orange, female-orange - 4 labels
-    MULTILABEL_CLASSIFICARION = "multilabel_classification" # e.g. male/female and the other blue/orange - 2 labels
-    OBJECT_DETECTION = "object_detection" # ?
-    SEMANTIC_SEGMENTATION = "semantic_segmentation" # ?
-    INSTANCE_SEGMENTATION = "instance segmentation"
-    KEYPOINT_DETECTION = "keypoint_detection"
-    MISC = "misc"
-    RAW = "raw"
-
+from enums import *
 
 # ---- PYDANTIC CLASSES ----
