+++ src/luxonis_ml/scheme/v1/enums/__init__.py
@@ -1,0 +1,6 @@
+from .data_type import DataType
+from .input_type import InputType
+from .platform import Platform
+from .postprocessing_block_type import PostprocessingBlockType
+from .preprocessing_block_type import PreprocessingBlockType
+from task_type import TaskType
