+++ src/luxonis_ml/scheme/v1/enums/data_type.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+class DataType(Enum):
+
+    """ all existing i/o data types"""
+
+    INT8 = "int8"
+    UINT8 = "uint8"
+    FLOAT32 = "float32"
