+++ src/luxonis_ml/scheme/v1/enums/input_type.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class InputType(Enum):
+
+    """ what type of input the model is expecting """
+
+    RAW = "raw"
+    IMAGE = "image"
