+++ src/luxonis_ml/scheme/v1/enums/platform.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+class Platform(Enum):
+    """ all existing Luxonis platforms """
+    HAILO = "hailo"
+    RVC2 = "rvc2"
+    RVC3 = "rvc3"
+    RVC4 = "rvc4"
