+++ src/luxonis_ml/scheme/v1/enums/postprocessing_block_type
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class PostprocessingBlockType(Enum):
+
+    """ block used in postprocessing """
+
+    MISC = "misc" # TODO
