+++ src/luxonis_ml/scheme/v1/enums/preprocessing_block_type
@@ -1,0 +1,7 @@
+class PreprocessingBlockType(Enum): # required only for RVC4
+
+    """ block used in preprocessing """
+
+    MEAN = "mean"
+    SCALE = "scale"
+    REVERSE_CHANNELS = "reverse_channels" #bgr<->rgb
